Correcting bugs
Need to send in the request to authenticate, otherwise the functions break.

If statement also corrected. Was relying on the err to be caught to return a 401 (email/pass dont match), however the err is only caught when the request doesn't go through, so in the case we have, the request go through, it catches the bad login attempt